Common metrics package
